fix: перенести валидацию из сервера на фасадный слой fix: добавить аргумент логеру
