+++ gateway/src/main/java/ru/practicum/shareit/exceptionhandler/ErrorHandler.java
@@ -6,6 +6,7 @@ import org.springframework.web.bind.MethodArgumentNotValidException;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestControllerAdvice;
+import ru.practicum.shareit.exceptionhandler.exceptions.DateTimeBookingException;
 
 @RestControllerAdvice("ru.practicum.shareit")
 @Slf4j
@@ -22,6 +23,13 @@ public class ErrorHandler {
     @ResponseStatus(HttpStatus.BAD_REQUEST)
     public ErrorResponse validationExceptionHandle(MethodArgumentNotValidException exception) {
         log.warn("Данные не прошли валидацию \n {}", exception);
+        return new ErrorResponse(exception.getMessage());
+    }
+
+    @ExceptionHandler
+    @ResponseStatus(HttpStatus.BAD_REQUEST)
+    public ErrorResponse dateTimeWrongParameterHandle(DateTimeBookingException exception) {
+        log.warn("Параметры для запроса на бронирование введены неверно. \n {}", exception);
         return new ErrorResponse(exception.getMessage());
     }
 }
