+++ server/src/main/resources/application.properties
@@ -2,6 +2,7 @@ spring.jpa.hibernate.ddl-auto=none
 spring.jpa.properties.hibernate.dialect=org.hibernate.dialect.PostgreSQL10Dialect
 spring.jpa.properties.hibernate.format_sql=true
 spring.sql.init.mode=always
+spring.jpa.defer-datasource-initialization=true
 spring.jpa.show-sql=true
 
 logging.level.org.springframework.orm.jpa=INFO
