+++ server/src/test/java/ru/practicum/shareit/test/services/BookingServiceImplTest.java
@@ -82,87 +82,6 @@ class BookingServiceImplTest {
     }
 
     @Test
-    void testAddNewRequest_ShouldReturnError_WhenBookingRequestDtoStartDateNull() {
-        // given
-        BookingRequestDto bookingRequestDto = makeDefaultBookingRequestDto();
-        bookingRequestDto.setStart(null);
-
-        // expect
-        final DateTimeBookingException exception = Assertions.assertThrows(
-                DateTimeBookingException.class,
-                () -> bookingService.addNewRequest(1L, bookingRequestDto));
-
-        assertThat(exception.getMessage(),
-                equalTo("Значения начала бронирования и конца бронирования должны быть указаны"));
-    }
-
-    @Test
-    void testAddNewRequest_ShouldReturnError_WhenBookingRequestDtoEndDateNull() {
-        // given
-        BookingRequestDto bookingRequestDto = makeDefaultBookingRequestDto();
-        bookingRequestDto.setEnd(null);
-
-        // expect
-        final DateTimeBookingException exception = Assertions.assertThrows(
-                DateTimeBookingException.class,
-                () -> bookingService.addNewRequest(1L, bookingRequestDto));
-
-        assertThat(exception.getMessage(),
-                equalTo("Значения начала бронирования и конца бронирования должны быть указаны"));
-    }
-
-    @Test
-    void testAddNewRequest_ShouldReturnError_WhenBookingRequestDtoEndDateAndStartDateSame() {
-        // given
-        BookingRequestDto bookingRequestDto = makeDefaultBookingRequestDto();
-        LocalDateTime plusFiveHourseFromNow = LocalDateTime.now().plusHours(5);
-        bookingRequestDto.setEnd(plusFiveHourseFromNow);
-        bookingRequestDto.setStart(plusFiveHourseFromNow);
-
-        // expect
-        final DateTimeBookingException exception = Assertions.assertThrows(
-                DateTimeBookingException.class,
-                () -> bookingService.addNewRequest(1L, bookingRequestDto));
-
-        assertThat(exception.getMessage(),
-                equalTo("Дата и время начала и конца бронирования не должны быть равны"));
-    }
-
-    @Test
-    void testAddNewRequest_ShouldReturnError_WhenBookingRequestDtoStartDateIsAfterEndDate() {
-        // given
-        BookingRequestDto bookingRequestDto = makeDefaultBookingRequestDto();
-        bookingRequestDto.setEnd(LocalDateTime.now().plusHours(1));
-        bookingRequestDto.setStart(LocalDateTime.now().plusHours(5));
-
-        // expect
-        final DateTimeBookingException exception = Assertions.assertThrows(
-                DateTimeBookingException.class,
-                () -> bookingService.addNewRequest(1L, bookingRequestDto));
-
-        assertThat(exception.getMessage(),
-                equalTo("Дата и время начала бронирования не может быть после окончания: "
-                        + bookingRequestDto.getStart()));
-    }
-
-    @Test
-    void testAddNewRequest_ShouldReturnError_WhenBookingRequestDtoStartDateIsBeforeNow() {
-        // given
-        BookingRequestDto bookingRequestDto = makeDefaultBookingRequestDto();
-        bookingRequestDto.setEnd(LocalDateTime.now().plusHours(1));
-        bookingRequestDto.setStart(LocalDateTime.now().minusDays(1));
-
-        // expect
-        final DateTimeBookingException exception = Assertions.assertThrows(
-                DateTimeBookingException.class,
-                () -> bookingService.addNewRequest(1L, bookingRequestDto));
-
-        assertThat(exception.getMessage(),
-                equalTo("Дата и время бронирования не может быть в прошлом: "
-                        + bookingRequestDto.getStart()));
-    }
-
-    @Test
     void testAddNewRequest_ShouldReturnError_WhenItemIsNull() {
         // given
         BookingRequestDto bookingRequestDto = makeDefaultBookingRequestDto();
